123: rename to com.omen273.crossLingo
