+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.example.crosswordToLearn"
+        applicationId "com.omen273.crossLingo"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
