+++ app/src/androidTest/java/com/omen273/crossLingo/ChooseTopicsInstrumentedTest.kt
@@ -1,15 +1,13 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.IdlingPolicies
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.filters.FlakyTest
 
 
 import org.junit.Rule
