+++ app/src/androidTest/java/com/omen273/crossLingo/NotEnoughWordsBecauseTooLongInstrumentedTest.kt
@@ -1,19 +1,19 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class NotEnoughWordsInstrumentedTest : ChoseTopicsToastTest() {
+class NotEnoughWordsBecauseTooLongInstrumentedTest : ChoseTopicsToastTest() {
 
     @Test
-    fun notEnoughWordsInstrumentedTest() {
+    fun notEnoughWordsBecauseTooLongInstrumentedTest() {
         choseTopicsImpl(
-            "notEnoughWords.json", getContext().getString(
+                "tooLongWordsData.json", getContext().getString(
                 R.string.not_enough_words,
                 ChooseTopicsActivity.CROSSWORD_SIZE
-            )
+        )
         )
     }
 }
