+++ app/src/androidTest/java/com/omen273/crossLingo/NotEnoughWordsInstrumentedTest.kt
@@ -1,19 +1,19 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class NotEnoughWordsBecauseTooLongInstrumentedTest : ChoseTopicsToastTest() {
+class NotEnoughWordsInstrumentedTest : ChoseTopicsToastTest() {
 
     @Test
-    fun notEnoughWordsBecauseTooLongInstrumentedTest() {
+    fun notEnoughWordsInstrumentedTest() {
         choseTopicsImpl(
-                "tooLongWordsData.json", getContext().getString(
+            "notEnoughWords.json", getContext().getString(
                 R.string.not_enough_words,
                 ChooseTopicsActivity.CROSSWORD_SIZE
-        )
+            )
         )
     }
 }
