+++ app/src/androidTest/java/com/omen273/crossLingo/SolveCrosswordInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Test
