+++ app/src/androidTest/java/com/omen273/crossLingo/SolveCrosswordWithTipsInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 import androidx.test.espresso.Espresso.*
 import androidx.test.espresso.action.ViewActions.click
@@ -8,7 +8,7 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import com.example.crosswordToLearn.R.id
+import com.omen273.crossLingo.R.id
 import junit.framework.TestCase.assertEquals
 import org.akop.ararat.core.Crossword
 import org.hamcrest.core.AnyOf.anyOf
