+++ app/src/androidTest/java/com/omen273/crossLingo/Utils.kt
@@ -1,4 +1,4 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 import android.content.Context
 import android.content.Intent
