+++ app/src/main/java/com/omen273/crossLingo/ConfigReader.kt
@@ -1,4 +1,4 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 import android.util.JsonReader
 import java.io.InputStream
