+++ app/src/main/java/com/omen273/crossLingo/ConfigWriter.kt
@@ -1,4 +1,4 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 import android.util.JsonWriter
 import java.io.OutputStream
