+++ app/src/main/java/com/omen273/crossLingo/CrosswordBuilderWrapper.kt
@@ -1,4 +1,4 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 class WordParams{
     var x: Int = 0
