+++ app/src/test/java/com/omen273/crossLingo/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.crosswordToLearn
+package com.omen273.crossLingo
 
 import org.junit.Test
